requirements.txt
streamlit
pandas
numpy
tensorflow-cpu
wordcloud
matplotlib
plotly